requirements.txt
fastapi
uvicorn
PyPDF2
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
python-multipart
boto3
